+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/encode/bytes/ResponseBodyClient.java
@@ -108,7 +108,7 @@ public final class ResponseBodyClient {
      * 
      * <pre>
      * {@code
-     * byte[]
+     * Base64Uri
      * }
      * </pre>
      *
