+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/encode/bytes/implementation/ResponseBodiesImpl.java
@@ -11,6 +11,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
+import io.clientcore.core.utils.Base64Uri;
 
 /**
  * An instance of this class provides access to all the operations defined in ResponseBodies.
@@ -77,7 +78,8 @@ public final class ResponseBodiesImpl {
         @HttpRequestInformation(
             method = HttpMethod.GET,
             path = "/encode/bytes/body/response/base64url",
-            expectedStatusCodes = { 200 })
+            expectedStatusCodes = { 200 },
+            returnValueWireType = Base64Uri.class)
         @UnexpectedResponseExceptionDetail
         Response<byte[]> base64urlSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
@@ -165,7 +167,7 @@ public final class ResponseBodiesImpl {
      * 
      * <pre>
      * {@code
-     * byte[]
+     * Base64Uri
      * }
      * </pre>
      *
