+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/EnumsOnlyClient.java
@@ -7,7 +7,7 @@ import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
 import type.union.implementation.EnumsOnliesImpl;
-import type.union.implementation.SendRequest3;
+import type.union.implementation.SendRequest6;
 
 /**
  * Initializes a new instance of the synchronous UnionClient type.
@@ -47,7 +47,7 @@ public final class EnumsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse3> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse6> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -66,14 +66,14 @@ public final class EnumsOnlyClient {
      * }
      * </pre>
      * 
-     * @param sendRequest3 The sendRequest3 parameter.
+     * @param sendRequest6 The sendRequest6 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest3, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest3, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest6, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest6, requestOptions);
     }
 
     /**
@@ -84,7 +84,7 @@ public final class EnumsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse3 get() {
+    public GetResponse6 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -102,8 +102,8 @@ public final class EnumsOnlyClient {
     public void send(EnumsOnlyCases prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest3 sendRequest3Obj = new SendRequest3(prop);
-        BinaryData sendRequest3 = BinaryData.fromObject(sendRequest3Obj);
-        sendWithResponse(sendRequest3, requestOptions).getValue();
+        SendRequest6 sendRequest6Obj = new SendRequest6(prop);
+        BinaryData sendRequest6 = BinaryData.fromObject(sendRequest6Obj);
+        sendWithResponse(sendRequest6, requestOptions).getValue();
     }
 }
