+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse.java
@@ -17,7 +17,7 @@ public final class GetResponse implements JsonSerializable<GetResponse> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final MixedTypesCases prop;
+    private final GetResponseProp prop;
 
     /**
      * Creates an instance of GetResponse class.
@@ -25,7 +25,7 @@ public final class GetResponse implements JsonSerializable<GetResponse> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse(MixedTypesCases prop) {
+    private GetResponse(GetResponseProp prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse implements JsonSerializable<GetResponse> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public MixedTypesCases getProp() {
+    public GetResponseProp getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse implements JsonSerializable<GetResponse> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.toString());
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse implements JsonSerializable<GetResponse> {
     @Metadata(generated = true)
     public static GetResponse fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            MixedTypesCases prop = null;
+            GetResponseProp prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = MixedTypesCases.fromJson(reader);
+                    prop = GetResponseProp.fromString(reader.getString());
                 } else {
                     reader.skipChildren();
                 }
