+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse1.java
@@ -17,7 +17,7 @@ public final class GetResponse1 implements JsonSerializable<GetResponse1> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final MixedLiteralsCases prop;
+    private final GetResponseProp1 prop;
 
     /**
      * Creates an instance of GetResponse1 class.
@@ -25,7 +25,7 @@ public final class GetResponse1 implements JsonSerializable<GetResponse1> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse1(MixedLiteralsCases prop) {
+    private GetResponse1(GetResponseProp1 prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse1 implements JsonSerializable<GetResponse1> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public MixedLiteralsCases getProp() {
+    public GetResponseProp1 getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse1 implements JsonSerializable<GetResponse1> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse1 implements JsonSerializable<GetResponse1> {
     @Metadata(generated = true)
     public static GetResponse1 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            MixedLiteralsCases prop = null;
+            GetResponseProp1 prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = MixedLiteralsCases.fromJson(reader);
+                    prop = GetResponseProp1.fromValue(reader.getString());
                 } else {
                     reader.skipChildren();
                 }
