+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse2.java
@@ -17,7 +17,7 @@ public final class GetResponse2 implements JsonSerializable<GetResponse2> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final StringAndArrayCases prop;
+    private final StringExtensibleNamedUnion prop;
 
     /**
      * Creates an instance of GetResponse2 class.
@@ -25,7 +25,7 @@ public final class GetResponse2 implements JsonSerializable<GetResponse2> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse2(StringAndArrayCases prop) {
+    private GetResponse2(StringExtensibleNamedUnion prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse2 implements JsonSerializable<GetResponse2> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public StringAndArrayCases getProp() {
+    public StringExtensibleNamedUnion getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse2 implements JsonSerializable<GetResponse2> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse2 implements JsonSerializable<GetResponse2> {
     @Metadata(generated = true)
     public static GetResponse2 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            StringAndArrayCases prop = null;
+            StringExtensibleNamedUnion prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = StringAndArrayCases.fromJson(reader);
+                    prop = StringExtensibleNamedUnion.fromValue(reader.getString());
                 } else {
                     reader.skipChildren();
                 }
