+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse3.java
@@ -17,7 +17,7 @@ public final class GetResponse3 implements JsonSerializable<GetResponse3> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final EnumsOnlyCases prop;
+    private final GetResponseProp2 prop;
 
     /**
      * Creates an instance of GetResponse3 class.
@@ -25,7 +25,7 @@ public final class GetResponse3 implements JsonSerializable<GetResponse3> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse3(EnumsOnlyCases prop) {
+    private GetResponse3(GetResponseProp2 prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse3 implements JsonSerializable<GetResponse3> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public EnumsOnlyCases getProp() {
+    public GetResponseProp2 getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse3 implements JsonSerializable<GetResponse3> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeNumberField("prop", this.prop == null ? null : this.prop.toInt());
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse3 implements JsonSerializable<GetResponse3> {
     @Metadata(generated = true)
     public static GetResponse3 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            EnumsOnlyCases prop = null;
+            GetResponseProp2 prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = EnumsOnlyCases.fromJson(reader);
+                    prop = GetResponseProp2.fromInt(reader.getInt());
                 } else {
                     reader.skipChildren();
                 }
