+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse5.java
@@ -2,6 +2,7 @@ package type.union;
 
 import io.clientcore.core.annotations.Metadata;
 import io.clientcore.core.annotations.TypeConditions;
+import io.clientcore.core.models.binarydata.BinaryData;
 import io.clientcore.core.serialization.json.JsonReader;
 import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
@@ -17,7 +18,7 @@ public final class GetResponse5 implements JsonSerializable<GetResponse5> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp1 prop;
+    private final BinaryData prop;
 
     /**
      * Creates an instance of GetResponse5 class.
@@ -25,7 +26,7 @@ public final class GetResponse5 implements JsonSerializable<GetResponse5> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse5(GetResponseProp1 prop) {
+    private GetResponse5(BinaryData prop) {
         this.prop = prop;
     }
 
@@ -35,7 +36,7 @@ public final class GetResponse5 implements JsonSerializable<GetResponse5> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp1 getProp() {
+    public BinaryData getProp() {
         return this.prop;
     }
 
@@ -46,7 +47,8 @@ public final class GetResponse5 implements JsonSerializable<GetResponse5> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeNumberField("prop", this.prop == null ? null : this.prop.toDouble());
+        jsonWriter.writeFieldName("prop");
+        this.prop.writeTo(jsonWriter);
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +64,13 @@ public final class GetResponse5 implements JsonSerializable<GetResponse5> {
     @Metadata(generated = true)
     public static GetResponse5 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp1 prop = null;
+            BinaryData prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp1.fromDouble(reader.getDouble());
+                    prop = reader.getNullable(nonNullReader -> BinaryData.fromObject(nonNullReader.readUntyped()));
                 } else {
                     reader.skipChildren();
                 }
