+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse6.java
@@ -17,7 +17,7 @@ public final class GetResponse6 implements JsonSerializable<GetResponse6> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp2 prop;
+    private final EnumsOnlyCases prop;
 
     /**
      * Creates an instance of GetResponse6 class.
@@ -25,7 +25,7 @@ public final class GetResponse6 implements JsonSerializable<GetResponse6> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse6(GetResponseProp2 prop) {
+    private GetResponse6(EnumsOnlyCases prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse6 implements JsonSerializable<GetResponse6> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp2 getProp() {
+    public EnumsOnlyCases getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse6 implements JsonSerializable<GetResponse6> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeNumberField("prop", this.prop == null ? null : this.prop.toInt());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse6 implements JsonSerializable<GetResponse6> {
     @Metadata(generated = true)
     public static GetResponse6 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp2 prop = null;
+            EnumsOnlyCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp2.fromInt(reader.getInt());
+                    prop = EnumsOnlyCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
