+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse7.java
@@ -17,7 +17,7 @@ public final class GetResponse7 implements JsonSerializable<GetResponse7> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final StringExtensibleNamedUnion prop;
+    private final StringAndArrayCases prop;
 
     /**
      * Creates an instance of GetResponse7 class.
@@ -25,7 +25,7 @@ public final class GetResponse7 implements JsonSerializable<GetResponse7> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse7(StringExtensibleNamedUnion prop) {
+    private GetResponse7(StringAndArrayCases prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse7 implements JsonSerializable<GetResponse7> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public StringExtensibleNamedUnion getProp() {
+    public StringAndArrayCases getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse7 implements JsonSerializable<GetResponse7> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse7 implements JsonSerializable<GetResponse7> {
     @Metadata(generated = true)
     public static GetResponse7 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            StringExtensibleNamedUnion prop = null;
+            StringAndArrayCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = StringExtensibleNamedUnion.fromValue(reader.getString());
+                    prop = StringAndArrayCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
