+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse8.java
@@ -17,7 +17,7 @@ public final class GetResponse8 implements JsonSerializable<GetResponse8> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp3 prop;
+    private final MixedLiteralsCases prop;
 
     /**
      * Creates an instance of GetResponse8 class.
@@ -25,7 +25,7 @@ public final class GetResponse8 implements JsonSerializable<GetResponse8> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse8(GetResponseProp3 prop) {
+    private GetResponse8(MixedLiteralsCases prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse8 implements JsonSerializable<GetResponse8> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp3 getProp() {
+    public MixedLiteralsCases getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse8 implements JsonSerializable<GetResponse8> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse8 implements JsonSerializable<GetResponse8> {
     @Metadata(generated = true)
     public static GetResponse8 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp3 prop = null;
+            MixedLiteralsCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp3.fromValue(reader.getString());
+                    prop = MixedLiteralsCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
