+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponse9.java
@@ -17,7 +17,7 @@ public final class GetResponse9 implements JsonSerializable<GetResponse9> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp4 prop;
+    private final MixedTypesCases prop;
 
     /**
      * Creates an instance of GetResponse9 class.
@@ -25,7 +25,7 @@ public final class GetResponse9 implements JsonSerializable<GetResponse9> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    private GetResponse9(GetResponseProp4 prop) {
+    private GetResponse9(MixedTypesCases prop) {
         this.prop = prop;
     }
 
@@ -35,7 +35,7 @@ public final class GetResponse9 implements JsonSerializable<GetResponse9> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp4 getProp() {
+    public MixedTypesCases getProp() {
         return this.prop;
     }
 
@@ -46,7 +46,7 @@ public final class GetResponse9 implements JsonSerializable<GetResponse9> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.toString());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -62,13 +62,13 @@ public final class GetResponse9 implements JsonSerializable<GetResponse9> {
     @Metadata(generated = true)
     public static GetResponse9 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp4 prop = null;
+            MixedTypesCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp4.fromString(reader.getString());
+                    prop = MixedTypesCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
