+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/GetResponseProp1.java
@@ -1,55 +1,127 @@
 package type.union;
 
+import io.clientcore.core.annotations.Metadata;
+import io.clientcore.core.serialization.json.JsonReader;
+import io.clientcore.core.serialization.json.JsonSerializable;
+import io.clientcore.core.serialization.json.JsonToken;
+import io.clientcore.core.serialization.json.JsonWriter;
+import io.clientcore.core.utils.ExpandableEnum;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Function;
+
 /**
  * Defines values for GetResponseProp1.
  */
-public enum GetResponseProp1 {
-    /**
-     * Enum value 1.1.
-     */
-    ONE_ONE(1.1),
+public final class GetResponseProp1 implements ExpandableEnum<String>, JsonSerializable<GetResponseProp1> {
+    private static final Map<String, GetResponseProp1> VALUES = new ConcurrentHashMap<>();
+
+    private static final Function<String, GetResponseProp1> NEW_INSTANCE = GetResponseProp1::new;
 
     /**
-     * Enum value 2.2.
+     * Static value b for GetResponseProp1.
      */
-    TWO_TWO(2.2),
+    @Metadata(generated = true)
+    public static final GetResponseProp1 B = fromValue("b");
 
     /**
-     * Enum value 3.3.
+     * Static value c for GetResponseProp1.
      */
-    THREE_THREE(3.3);
+    @Metadata(generated = true)
+    public static final GetResponseProp1 C = fromValue("c");
 
-    /**
-     * The actual serialized value for a GetResponseProp1 instance.
-     */
-    private final double value;
+    private final String value;
 
-    GetResponseProp1(double value) {
+    private GetResponseProp1(String value) {
         this.value = value;
     }
 
     /**
-     * Parses a serialized value to a GetResponseProp1 instance.
+     * Creates or finds a GetResponseProp1.
      * 
-     * @param value the serialized value to parse.
-     * @return the parsed GetResponseProp1 object, or null if unable to parse.
+     * @param value a value to look for.
+     * @return the corresponding GetResponseProp1.
+     * @throws IllegalArgumentException if value is null.
      */
-    public static GetResponseProp1 fromDouble(double value) {
-        GetResponseProp1[] items = GetResponseProp1.values();
-        for (GetResponseProp1 item : items) {
-            if (Double.doubleToLongBits(item.toDouble()) == Double.doubleToLongBits(value)) {
-                return item;
-            }
+    @Metadata(generated = true)
+    public static GetResponseProp1 fromValue(String value) {
+        if (value == null) {
+            throw new IllegalArgumentException("'value' cannot be null.");
         }
-        return null;
+        return VALUES.computeIfAbsent(value, NEW_INSTANCE);
     }
 
     /**
-     * De-serializes the instance to double value.
+     * Gets known GetResponseProp1 values.
      * 
-     * @return the double value.
+     * @return Known GetResponseProp1 values.
      */
-    public double toDouble() {
+    @Metadata(generated = true)
+    public static Collection<GetResponseProp1> values() {
+        return new ArrayList<>(VALUES.values());
+    }
+
+    /**
+     * Gets the value of the GetResponseProp1 instance.
+     * 
+     * @return the value of the GetResponseProp1 instance.
+     */
+    @Metadata(generated = true)
+    @Override
+    public String getValue() {
         return this.value;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Metadata(generated = true)
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        return jsonWriter.writeString(getValue());
+    }
+
+    /**
+     * Reads an instance of GetResponseProp1 from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of GetResponseProp1 if the JsonReader was pointing to an instance of it, or null if the
+     * JsonReader was pointing to JSON null.
+     * @throws IOException If an error occurs while reading the GetResponseProp1.
+     * @throws IllegalStateException If unexpected JSON token is found.
+     */
+    @Metadata(generated = true)
+    public static GetResponseProp1 fromJson(JsonReader jsonReader) throws IOException {
+        JsonToken nextToken = jsonReader.nextToken();
+        if (nextToken == JsonToken.NULL) {
+            return null;
+        }
+        if (nextToken != JsonToken.STRING) {
+            throw new IllegalStateException(
+                String.format("Unexpected JSON token for %s deserialization: %s", JsonToken.STRING, nextToken));
+        }
+        return GetResponseProp1.fromValue(jsonReader.getString());
+    }
+
+    @Metadata(generated = true)
+    @Override
+    public String toString() {
+        return Objects.toString(this.value);
+    }
+
+    @Metadata(generated = true)
+    @Override
+    public boolean equals(Object obj) {
+        return this == obj;
+    }
+
+    @Metadata(generated = true)
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(this.value);
     }
 }
