+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/IntsOnlyClient.java
@@ -7,7 +7,7 @@ import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
 import type.union.implementation.IntsOnliesImpl;
-import type.union.implementation.SendRequest6;
+import type.union.implementation.SendRequest3;
 
 /**
  * Initializes a new instance of the synchronous UnionClient type.
@@ -44,7 +44,7 @@ public final class IntsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse6> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse3> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -60,14 +60,14 @@ public final class IntsOnlyClient {
      * }
      * </pre>
      * 
-     * @param sendRequest6 The sendRequest6 parameter.
+     * @param sendRequest3 The sendRequest3 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest6, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest6, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest3, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest3, requestOptions);
     }
 
     /**
@@ -78,7 +78,7 @@ public final class IntsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse6 get() {
+    public GetResponse3 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -96,8 +96,8 @@ public final class IntsOnlyClient {
     public void send(GetResponseProp2 prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest6 sendRequest6Obj = new SendRequest6(prop);
-        BinaryData sendRequest6 = BinaryData.fromObject(sendRequest6Obj);
-        sendWithResponse(sendRequest6, requestOptions).getValue();
+        SendRequest3 sendRequest3Obj = new SendRequest3(prop);
+        BinaryData sendRequest3 = BinaryData.fromObject(sendRequest3Obj);
+        sendWithResponse(sendRequest3, requestOptions).getValue();
     }
 }
