+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/MixedLiteralsClient.java
@@ -7,7 +7,7 @@ import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
 import type.union.implementation.MixedLiteralsImpl;
-import type.union.implementation.SendRequest1;
+import type.union.implementation.SendRequest8;
 
 /**
  * Initializes a new instance of the synchronous UnionClient type.
@@ -49,7 +49,7 @@ public final class MixedLiteralsClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse1> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse8> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -70,14 +70,14 @@ public final class MixedLiteralsClient {
      * }
      * </pre>
      * 
-     * @param sendRequest1 The sendRequest1 parameter.
+     * @param sendRequest8 The sendRequest8 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest1, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest1, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest8, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest8, requestOptions);
     }
 
     /**
@@ -88,7 +88,7 @@ public final class MixedLiteralsClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse1 get() {
+    public GetResponse8 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -106,8 +106,8 @@ public final class MixedLiteralsClient {
     public void send(MixedLiteralsCases prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest1 sendRequest1Obj = new SendRequest1(prop);
-        BinaryData sendRequest1 = BinaryData.fromObject(sendRequest1Obj);
-        sendWithResponse(sendRequest1, requestOptions).getValue();
+        SendRequest8 sendRequest8Obj = new SendRequest8(prop);
+        BinaryData sendRequest8 = BinaryData.fromObject(sendRequest8Obj);
+        sendWithResponse(sendRequest8, requestOptions).getValue();
     }
 }
