+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/MixedTypesClient.java
@@ -7,7 +7,7 @@ import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
 import type.union.implementation.MixedTypesImpl;
-import type.union.implementation.SendRequest;
+import type.union.implementation.SendRequest9;
 
 /**
  * Initializes a new instance of the synchronous UnionClient type.
@@ -52,7 +52,7 @@ public final class MixedTypesClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse9> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -76,14 +76,14 @@ public final class MixedTypesClient {
      * }
      * </pre>
      * 
-     * @param sendRequest The sendRequest parameter.
+     * @param sendRequest9 The sendRequest9 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest9, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest9, requestOptions);
     }
 
     /**
@@ -94,7 +94,7 @@ public final class MixedTypesClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse get() {
+    public GetResponse9 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -112,8 +112,8 @@ public final class MixedTypesClient {
     public void send(MixedTypesCases prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest sendRequestObj = new SendRequest(prop);
-        BinaryData sendRequest = BinaryData.fromObject(sendRequestObj);
-        sendWithResponse(sendRequest, requestOptions).getValue();
+        SendRequest9 sendRequest9Obj = new SendRequest9(prop);
+        BinaryData sendRequest9 = BinaryData.fromObject(sendRequest9Obj);
+        sendWithResponse(sendRequest9, requestOptions).getValue();
     }
 }
