+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/ModelsOnlyClient.java
@@ -7,7 +7,7 @@ import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
 import type.union.implementation.ModelsOnliesImpl;
-import type.union.implementation.SendRequest4;
+import type.union.implementation.SendRequest5;
 
 /**
  * Initializes a new instance of the synchronous UnionClient type.
@@ -44,7 +44,7 @@ public final class ModelsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse4> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse5> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -60,14 +60,14 @@ public final class ModelsOnlyClient {
      * }
      * </pre>
      * 
-     * @param sendRequest4 The sendRequest4 parameter.
+     * @param sendRequest5 The sendRequest5 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest4, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest4, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest5, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest5, requestOptions);
     }
 
     /**
@@ -78,7 +78,7 @@ public final class ModelsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse4 get() {
+    public GetResponse5 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -96,8 +96,8 @@ public final class ModelsOnlyClient {
     public void send(BinaryData prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest4 sendRequest4Obj = new SendRequest4(prop);
-        BinaryData sendRequest4 = BinaryData.fromObject(sendRequest4Obj);
-        sendWithResponse(sendRequest4, requestOptions).getValue();
+        SendRequest5 sendRequest5Obj = new SendRequest5(prop);
+        BinaryData sendRequest5 = BinaryData.fromObject(sendRequest5Obj);
+        sendWithResponse(sendRequest5, requestOptions).getValue();
     }
 }
