+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/StringExtensibleClient.java
@@ -6,7 +6,7 @@ import io.clientcore.core.http.exceptions.HttpResponseException;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.implementation.SendRequest8;
+import type.union.implementation.SendRequest1;
 import type.union.implementation.StringExtensiblesImpl;
 
 /**
@@ -44,7 +44,7 @@ public final class StringExtensibleClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse8> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse1> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -60,14 +60,14 @@ public final class StringExtensibleClient {
      * }
      * </pre>
      * 
-     * @param sendRequest8 The sendRequest8 parameter.
+     * @param sendRequest1 The sendRequest1 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest8, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest8, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest1, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest1, requestOptions);
     }
 
     /**
@@ -78,7 +78,7 @@ public final class StringExtensibleClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse8 get() {
+    public GetResponse1 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -93,11 +93,11 @@ public final class StringExtensibleClient {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     @Metadata(generated = true)
-    public void send(GetResponseProp3 prop) {
+    public void send(GetResponseProp1 prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest8 sendRequest8Obj = new SendRequest8(prop);
-        BinaryData sendRequest8 = BinaryData.fromObject(sendRequest8Obj);
-        sendWithResponse(sendRequest8, requestOptions).getValue();
+        SendRequest1 sendRequest1Obj = new SendRequest1(prop);
+        BinaryData sendRequest1 = BinaryData.fromObject(sendRequest1Obj);
+        sendWithResponse(sendRequest1, requestOptions).getValue();
     }
 }
