+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/StringExtensibleNamedClient.java
@@ -6,7 +6,7 @@ import io.clientcore.core.http.exceptions.HttpResponseException;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.implementation.SendRequest7;
+import type.union.implementation.SendRequest2;
 import type.union.implementation.StringExtensibleNamedsImpl;
 
 /**
@@ -44,7 +44,7 @@ public final class StringExtensibleNamedClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse7> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse2> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -60,14 +60,14 @@ public final class StringExtensibleNamedClient {
      * }
      * </pre>
      * 
-     * @param sendRequest7 The sendRequest7 parameter.
+     * @param sendRequest2 The sendRequest2 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest7, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest7, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest2, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest2, requestOptions);
     }
 
     /**
@@ -78,7 +78,7 @@ public final class StringExtensibleNamedClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse7 get() {
+    public GetResponse2 get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -96,8 +96,8 @@ public final class StringExtensibleNamedClient {
     public void send(StringExtensibleNamedUnion prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest7 sendRequest7Obj = new SendRequest7(prop);
-        BinaryData sendRequest7 = BinaryData.fromObject(sendRequest7Obj);
-        sendWithResponse(sendRequest7, requestOptions).getValue();
+        SendRequest2 sendRequest2Obj = new SendRequest2(prop);
+        BinaryData sendRequest2 = BinaryData.fromObject(sendRequest2Obj);
+        sendWithResponse(sendRequest2, requestOptions).getValue();
     }
 }
