+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/StringsOnlyClient.java
@@ -6,7 +6,7 @@ import io.clientcore.core.http.exceptions.HttpResponseException;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.implementation.SendRequest9;
+import type.union.implementation.SendRequest;
 import type.union.implementation.StringsOnliesImpl;
 
 /**
@@ -44,7 +44,7 @@ public final class StringsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<GetResponse9> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse> getWithResponse(RequestOptions requestOptions) {
         return this.serviceClient.getWithResponse(requestOptions);
     }
 
@@ -60,14 +60,14 @@ public final class StringsOnlyClient {
      * }
      * </pre>
      * 
-     * @param sendRequest9 The sendRequest9 parameter.
+     * @param sendRequest The sendRequest parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
     @Metadata(generated = true)
-    public Response<Void> sendWithResponse(BinaryData sendRequest9, RequestOptions requestOptions) {
-        return this.serviceClient.sendWithResponse(sendRequest9, requestOptions);
+    public Response<Void> sendWithResponse(BinaryData sendRequest, RequestOptions requestOptions) {
+        return this.serviceClient.sendWithResponse(sendRequest, requestOptions);
     }
 
     /**
@@ -78,7 +78,7 @@ public final class StringsOnlyClient {
      * @return the response.
      */
     @Metadata(generated = true)
-    public GetResponse9 get() {
+    public GetResponse get() {
         // Generated convenience method for getWithResponse
         RequestOptions requestOptions = new RequestOptions();
         return getWithResponse(requestOptions).getValue();
@@ -93,11 +93,11 @@ public final class StringsOnlyClient {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     @Metadata(generated = true)
-    public void send(GetResponseProp4 prop) {
+    public void send(GetResponseProp prop) {
         // Generated convenience method for sendWithResponse
         RequestOptions requestOptions = new RequestOptions();
-        SendRequest9 sendRequest9Obj = new SendRequest9(prop);
-        BinaryData sendRequest9 = BinaryData.fromObject(sendRequest9Obj);
-        sendWithResponse(sendRequest9, requestOptions).getValue();
+        SendRequest sendRequestObj = new SendRequest(prop);
+        BinaryData sendRequest = BinaryData.fromObject(sendRequestObj);
+        sendWithResponse(sendRequest, requestOptions).getValue();
     }
 }
