+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/EnumsOnliesImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse3;
+import type.union.GetResponse6;
 
 /**
  * An instance of this class provides access to all the operations defined in EnumsOnlies.
@@ -46,7 +46,7 @@ public final class EnumsOnliesImpl {
     public interface EnumsOnliesService {
         @HttpRequestInformation(method = HttpMethod.GET, path = "/type/union/enums-only", expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse3> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse6> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -55,7 +55,7 @@ public final class EnumsOnliesImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest3, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest6, RequestOptions requestOptions);
     }
 
     /**
@@ -77,7 +77,7 @@ public final class EnumsOnliesImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse3> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse6> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -97,13 +97,13 @@ public final class EnumsOnliesImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest3 The sendRequest3 parameter.
+     * @param sendRequest6 The sendRequest6 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest3, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest6, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest3, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest6, requestOptions);
     }
 }
