+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/IntsOnliesImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse6;
+import type.union.GetResponse3;
 
 /**
  * An instance of this class provides access to all the operations defined in IntsOnlies.
@@ -46,13 +46,13 @@ public final class IntsOnliesImpl {
     public interface IntsOnliesService {
         @HttpRequestInformation(method = HttpMethod.GET, path = "/type/union/ints-only", expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse6> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse3> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(method = HttpMethod.POST, path = "/type/union/ints-only", expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest6, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest3, RequestOptions requestOptions);
     }
 
     /**
@@ -71,7 +71,7 @@ public final class IntsOnliesImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse6> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse3> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -88,13 +88,13 @@ public final class IntsOnliesImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest6 The sendRequest6 parameter.
+     * @param sendRequest3 The sendRequest3 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest6, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest3, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest6, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest3, requestOptions);
     }
 }
