+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/MixedLiteralsImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse1;
+import type.union.GetResponse8;
 
 /**
  * An instance of this class provides access to all the operations defined in MixedLiterals.
@@ -49,7 +49,7 @@ public final class MixedLiteralsImpl {
             path = "/type/union/mixed-literals",
             expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse1> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse8> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -58,7 +58,7 @@ public final class MixedLiteralsImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest1, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest8, RequestOptions requestOptions);
     }
 
     /**
@@ -82,7 +82,7 @@ public final class MixedLiteralsImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse1> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse8> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -104,13 +104,13 @@ public final class MixedLiteralsImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest1 The sendRequest1 parameter.
+     * @param sendRequest8 The sendRequest8 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest1, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest8, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest1, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest8, requestOptions);
     }
 }
