+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/MixedTypesImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse;
+import type.union.GetResponse9;
 
 /**
  * An instance of this class provides access to all the operations defined in MixedTypes.
@@ -49,7 +49,7 @@ public final class MixedTypesImpl {
             path = "/type/union/mixed-types",
             expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse9> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -58,7 +58,7 @@ public final class MixedTypesImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest9, RequestOptions requestOptions);
     }
 
     /**
@@ -85,7 +85,7 @@ public final class MixedTypesImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse9> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -110,13 +110,13 @@ public final class MixedTypesImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest The sendRequest parameter.
+     * @param sendRequest9 The sendRequest9 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest9, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest9, requestOptions);
     }
 }
