+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/ModelsOnliesImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse4;
+import type.union.GetResponse5;
 
 /**
  * An instance of this class provides access to all the operations defined in ModelsOnlies.
@@ -49,7 +49,7 @@ public final class ModelsOnliesImpl {
             path = "/type/union/models-only",
             expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse4> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse5> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -58,7 +58,7 @@ public final class ModelsOnliesImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest4, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest5, RequestOptions requestOptions);
     }
 
     /**
@@ -77,7 +77,7 @@ public final class ModelsOnliesImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse4> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse5> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -94,13 +94,13 @@ public final class ModelsOnliesImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest4 The sendRequest4 parameter.
+     * @param sendRequest5 The sendRequest5 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest4, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest5, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest4, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest5, requestOptions);
     }
 }
