+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.MixedTypesCases;
+import type.union.GetResponseProp;
 
 /**
  * The SendRequest model.
@@ -18,7 +18,7 @@ public final class SendRequest implements JsonSerializable<SendRequest> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final MixedTypesCases prop;
+    private final GetResponseProp prop;
 
     /**
      * Creates an instance of SendRequest class.
@@ -26,7 +26,7 @@ public final class SendRequest implements JsonSerializable<SendRequest> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest(MixedTypesCases prop) {
+    public SendRequest(GetResponseProp prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest implements JsonSerializable<SendRequest> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public MixedTypesCases getProp() {
+    public GetResponseProp getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest implements JsonSerializable<SendRequest> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.toString());
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest implements JsonSerializable<SendRequest> {
     @Metadata(generated = true)
     public static SendRequest fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            MixedTypesCases prop = null;
+            GetResponseProp prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = MixedTypesCases.fromJson(reader);
+                    prop = GetResponseProp.fromString(reader.getString());
                 } else {
                     reader.skipChildren();
                 }
