+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest1.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.MixedLiteralsCases;
+import type.union.GetResponseProp1;
 
 /**
  * The SendRequest1 model.
@@ -18,7 +18,7 @@ public final class SendRequest1 implements JsonSerializable<SendRequest1> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final MixedLiteralsCases prop;
+    private final GetResponseProp1 prop;
 
     /**
      * Creates an instance of SendRequest1 class.
@@ -26,7 +26,7 @@ public final class SendRequest1 implements JsonSerializable<SendRequest1> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest1(MixedLiteralsCases prop) {
+    public SendRequest1(GetResponseProp1 prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest1 implements JsonSerializable<SendRequest1> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public MixedLiteralsCases getProp() {
+    public GetResponseProp1 getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest1 implements JsonSerializable<SendRequest1> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest1 implements JsonSerializable<SendRequest1> {
     @Metadata(generated = true)
     public static SendRequest1 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            MixedLiteralsCases prop = null;
+            GetResponseProp1 prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = MixedLiteralsCases.fromJson(reader);
+                    prop = GetResponseProp1.fromValue(reader.getString());
                 } else {
                     reader.skipChildren();
                 }
