+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest2.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.StringAndArrayCases;
+import type.union.StringExtensibleNamedUnion;
 
 /**
  * The SendRequest2 model.
@@ -18,7 +18,7 @@ public final class SendRequest2 implements JsonSerializable<SendRequest2> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final StringAndArrayCases prop;
+    private final StringExtensibleNamedUnion prop;
 
     /**
      * Creates an instance of SendRequest2 class.
@@ -26,7 +26,7 @@ public final class SendRequest2 implements JsonSerializable<SendRequest2> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest2(StringAndArrayCases prop) {
+    public SendRequest2(StringExtensibleNamedUnion prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest2 implements JsonSerializable<SendRequest2> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public StringAndArrayCases getProp() {
+    public StringExtensibleNamedUnion getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest2 implements JsonSerializable<SendRequest2> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeJsonField("prop", this.prop);
+        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest2 implements JsonSerializable<SendRequest2> {
     @Metadata(generated = true)
     public static SendRequest2 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            StringAndArrayCases prop = null;
+            StringExtensibleNamedUnion prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = StringAndArrayCases.fromJson(reader);
+                    prop = StringExtensibleNamedUnion.fromValue(reader.getString());
                 } else {
                     reader.skipChildren();
                 }
