+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest4.java
@@ -2,12 +2,12 @@ package type.union.implementation;
 
 import io.clientcore.core.annotations.Metadata;
 import io.clientcore.core.annotations.TypeConditions;
-import io.clientcore.core.models.binarydata.BinaryData;
 import io.clientcore.core.serialization.json.JsonReader;
 import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
+import type.union.GetResponseProp3;
 
 /**
  * The SendRequest4 model.
@@ -18,7 +18,7 @@ public final class SendRequest4 implements JsonSerializable<SendRequest4> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final BinaryData prop;
+    private final GetResponseProp3 prop;
 
     /**
      * Creates an instance of SendRequest4 class.
@@ -26,7 +26,7 @@ public final class SendRequest4 implements JsonSerializable<SendRequest4> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest4(BinaryData prop) {
+    public SendRequest4(GetResponseProp3 prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest4 implements JsonSerializable<SendRequest4> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public BinaryData getProp() {
+    public GetResponseProp3 getProp() {
         return this.prop;
     }
 
@@ -47,8 +47,7 @@ public final class SendRequest4 implements JsonSerializable<SendRequest4> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeFieldName("prop");
-        this.prop.writeTo(jsonWriter);
+        jsonWriter.writeNumberField("prop", this.prop == null ? null : this.prop.toDouble());
         return jsonWriter.writeEndObject();
     }
 
@@ -64,13 +63,13 @@ public final class SendRequest4 implements JsonSerializable<SendRequest4> {
     @Metadata(generated = true)
     public static SendRequest4 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            BinaryData prop = null;
+            GetResponseProp3 prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = reader.getNullable(nonNullReader -> BinaryData.fromObject(nonNullReader.readUntyped()));
+                    prop = GetResponseProp3.fromDouble(reader.getDouble());
                 } else {
                     reader.skipChildren();
                 }
