+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest5.java
@@ -2,12 +2,12 @@ package type.union.implementation;
 
 import io.clientcore.core.annotations.Metadata;
 import io.clientcore.core.annotations.TypeConditions;
+import io.clientcore.core.models.binarydata.BinaryData;
 import io.clientcore.core.serialization.json.JsonReader;
 import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.GetResponseProp1;
 
 /**
  * The SendRequest5 model.
@@ -18,7 +18,7 @@ public final class SendRequest5 implements JsonSerializable<SendRequest5> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp1 prop;
+    private final BinaryData prop;
 
     /**
      * Creates an instance of SendRequest5 class.
@@ -26,7 +26,7 @@ public final class SendRequest5 implements JsonSerializable<SendRequest5> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest5(GetResponseProp1 prop) {
+    public SendRequest5(BinaryData prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest5 implements JsonSerializable<SendRequest5> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp1 getProp() {
+    public BinaryData getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,8 @@ public final class SendRequest5 implements JsonSerializable<SendRequest5> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeNumberField("prop", this.prop == null ? null : this.prop.toDouble());
+        jsonWriter.writeFieldName("prop");
+        this.prop.writeTo(jsonWriter);
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +64,13 @@ public final class SendRequest5 implements JsonSerializable<SendRequest5> {
     @Metadata(generated = true)
     public static SendRequest5 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp1 prop = null;
+            BinaryData prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp1.fromDouble(reader.getDouble());
+                    prop = reader.getNullable(nonNullReader -> BinaryData.fromObject(nonNullReader.readUntyped()));
                 } else {
                     reader.skipChildren();
                 }
