+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest6.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.GetResponseProp2;
+import type.union.EnumsOnlyCases;
 
 /**
  * The SendRequest6 model.
@@ -18,7 +18,7 @@ public final class SendRequest6 implements JsonSerializable<SendRequest6> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp2 prop;
+    private final EnumsOnlyCases prop;
 
     /**
      * Creates an instance of SendRequest6 class.
@@ -26,7 +26,7 @@ public final class SendRequest6 implements JsonSerializable<SendRequest6> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest6(GetResponseProp2 prop) {
+    public SendRequest6(EnumsOnlyCases prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest6 implements JsonSerializable<SendRequest6> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp2 getProp() {
+    public EnumsOnlyCases getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest6 implements JsonSerializable<SendRequest6> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeNumberField("prop", this.prop == null ? null : this.prop.toInt());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest6 implements JsonSerializable<SendRequest6> {
     @Metadata(generated = true)
     public static SendRequest6 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp2 prop = null;
+            EnumsOnlyCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp2.fromInt(reader.getInt());
+                    prop = EnumsOnlyCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
