+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest7.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.StringExtensibleNamedUnion;
+import type.union.StringAndArrayCases;
 
 /**
  * The SendRequest7 model.
@@ -18,7 +18,7 @@ public final class SendRequest7 implements JsonSerializable<SendRequest7> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final StringExtensibleNamedUnion prop;
+    private final StringAndArrayCases prop;
 
     /**
      * Creates an instance of SendRequest7 class.
@@ -26,7 +26,7 @@ public final class SendRequest7 implements JsonSerializable<SendRequest7> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest7(StringExtensibleNamedUnion prop) {
+    public SendRequest7(StringAndArrayCases prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest7 implements JsonSerializable<SendRequest7> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public StringExtensibleNamedUnion getProp() {
+    public StringAndArrayCases getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest7 implements JsonSerializable<SendRequest7> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest7 implements JsonSerializable<SendRequest7> {
     @Metadata(generated = true)
     public static SendRequest7 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            StringExtensibleNamedUnion prop = null;
+            StringAndArrayCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = StringExtensibleNamedUnion.fromValue(reader.getString());
+                    prop = StringAndArrayCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
