+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest8.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.GetResponseProp3;
+import type.union.MixedLiteralsCases;
 
 /**
  * The SendRequest8 model.
@@ -18,7 +18,7 @@ public final class SendRequest8 implements JsonSerializable<SendRequest8> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp3 prop;
+    private final MixedLiteralsCases prop;
 
     /**
      * Creates an instance of SendRequest8 class.
@@ -26,7 +26,7 @@ public final class SendRequest8 implements JsonSerializable<SendRequest8> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest8(GetResponseProp3 prop) {
+    public SendRequest8(MixedLiteralsCases prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest8 implements JsonSerializable<SendRequest8> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp3 getProp() {
+    public MixedLiteralsCases getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest8 implements JsonSerializable<SendRequest8> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.getValue());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest8 implements JsonSerializable<SendRequest8> {
     @Metadata(generated = true)
     public static SendRequest8 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp3 prop = null;
+            MixedLiteralsCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp3.fromValue(reader.getString());
+                    prop = MixedLiteralsCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
