+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/SendRequest9.java
@@ -7,7 +7,7 @@ import io.clientcore.core.serialization.json.JsonSerializable;
 import io.clientcore.core.serialization.json.JsonToken;
 import io.clientcore.core.serialization.json.JsonWriter;
 import java.io.IOException;
-import type.union.GetResponseProp4;
+import type.union.MixedTypesCases;
 
 /**
  * The SendRequest9 model.
@@ -18,7 +18,7 @@ public final class SendRequest9 implements JsonSerializable<SendRequest9> {
      * The prop property.
      */
     @Metadata(generated = true)
-    private final GetResponseProp4 prop;
+    private final MixedTypesCases prop;
 
     /**
      * Creates an instance of SendRequest9 class.
@@ -26,7 +26,7 @@ public final class SendRequest9 implements JsonSerializable<SendRequest9> {
      * @param prop the prop value to set.
      */
     @Metadata(generated = true)
-    public SendRequest9(GetResponseProp4 prop) {
+    public SendRequest9(MixedTypesCases prop) {
         this.prop = prop;
     }
 
@@ -36,7 +36,7 @@ public final class SendRequest9 implements JsonSerializable<SendRequest9> {
      * @return the prop value.
      */
     @Metadata(generated = true)
-    public GetResponseProp4 getProp() {
+    public MixedTypesCases getProp() {
         return this.prop;
     }
 
@@ -47,7 +47,7 @@ public final class SendRequest9 implements JsonSerializable<SendRequest9> {
     @Override
     public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
         jsonWriter.writeStartObject();
-        jsonWriter.writeStringField("prop", this.prop == null ? null : this.prop.toString());
+        jsonWriter.writeJsonField("prop", this.prop);
         return jsonWriter.writeEndObject();
     }
 
@@ -63,13 +63,13 @@ public final class SendRequest9 implements JsonSerializable<SendRequest9> {
     @Metadata(generated = true)
     public static SendRequest9 fromJson(JsonReader jsonReader) throws IOException {
         return jsonReader.readObject(reader -> {
-            GetResponseProp4 prop = null;
+            MixedTypesCases prop = null;
             while (reader.nextToken() != JsonToken.END_OBJECT) {
                 String fieldName = reader.getFieldName();
                 reader.nextToken();
 
                 if ("prop".equals(fieldName)) {
-                    prop = GetResponseProp4.fromString(reader.getString());
+                    prop = MixedTypesCases.fromJson(reader);
                 } else {
                     reader.skipChildren();
                 }
