+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/StringAndArraysImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse2;
+import type.union.GetResponse7;
 
 /**
  * An instance of this class provides access to all the operations defined in StringAndArrays.
@@ -49,7 +49,7 @@ public final class StringAndArraysImpl {
             path = "/type/union/string-and-array",
             expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse2> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse7> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -58,7 +58,7 @@ public final class StringAndArraysImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest2, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest7, RequestOptions requestOptions);
     }
 
     /**
@@ -80,7 +80,7 @@ public final class StringAndArraysImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse2> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse7> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -100,13 +100,13 @@ public final class StringAndArraysImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest2 The sendRequest2 parameter.
+     * @param sendRequest7 The sendRequest7 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest2, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest7, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest2, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest7, requestOptions);
     }
 }
