+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/StringExtensiblesImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse8;
+import type.union.GetResponse1;
 
 /**
  * An instance of this class provides access to all the operations defined in StringExtensibles.
@@ -49,7 +49,7 @@ public final class StringExtensiblesImpl {
             path = "/type/union/string-extensible",
             expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse8> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse1> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -58,7 +58,7 @@ public final class StringExtensiblesImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest8, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest1, RequestOptions requestOptions);
     }
 
     /**
@@ -77,7 +77,7 @@ public final class StringExtensiblesImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse8> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse1> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -94,13 +94,13 @@ public final class StringExtensiblesImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest8 The sendRequest8 parameter.
+     * @param sendRequest1 The sendRequest1 parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest8, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest1, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest8, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest1, requestOptions);
     }
 }
