+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/java/type/union/implementation/StringsOnliesImpl.java
@@ -12,7 +12,7 @@ import io.clientcore.core.http.models.HttpMethod;
 import io.clientcore.core.http.models.RequestOptions;
 import io.clientcore.core.http.models.Response;
 import io.clientcore.core.models.binarydata.BinaryData;
-import type.union.GetResponse9;
+import type.union.GetResponse;
 
 /**
  * An instance of this class provides access to all the operations defined in StringsOnlies.
@@ -49,7 +49,7 @@ public final class StringsOnliesImpl {
             path = "/type/union/strings-only",
             expectedStatusCodes = { 200 })
         @UnexpectedResponseExceptionDetail
-        Response<GetResponse9> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
+        Response<GetResponse> getSync(@HostParam("endpoint") String endpoint, @HeaderParam("Accept") String accept,
             RequestOptions requestOptions);
 
         @HttpRequestInformation(
@@ -58,7 +58,7 @@ public final class StringsOnliesImpl {
             expectedStatusCodes = { 204 })
         @UnexpectedResponseExceptionDetail
         Response<Void> sendSync(@HostParam("endpoint") String endpoint, @HeaderParam("Content-Type") String contentType,
-            @BodyParam("application/json") BinaryData sendRequest9, RequestOptions requestOptions);
+            @BodyParam("application/json") BinaryData sendRequest, RequestOptions requestOptions);
     }
 
     /**
@@ -77,7 +77,7 @@ public final class StringsOnliesImpl {
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<GetResponse9> getWithResponse(RequestOptions requestOptions) {
+    public Response<GetResponse> getWithResponse(RequestOptions requestOptions) {
         final String accept = "application/json";
         return service.getSync(this.client.getEndpoint(), accept, requestOptions);
     }
@@ -94,13 +94,13 @@ public final class StringsOnliesImpl {
      * }
      * </pre>
      * 
-     * @param sendRequest9 The sendRequest9 parameter.
+     * @param sendRequest The sendRequest parameter.
      * @param requestOptions The options to configure the HTTP request before HTTP client sends it.
      * @throws HttpResponseException thrown if the service returns an error.
      * @return the response.
      */
-    public Response<Void> sendWithResponse(BinaryData sendRequest9, RequestOptions requestOptions) {
+    public Response<Void> sendWithResponse(BinaryData sendRequest, RequestOptions requestOptions) {
         final String contentType = "application/json";
-        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest9, requestOptions);
+        return service.sendSync(this.client.getEndpoint(), contentType, sendRequest, requestOptions);
     }
 }
