+++ packages/http-client-java/generator/http-client-generator-clientcore-test/src/main/resources/META-INF/type-union_apiview_properties.json
@@ -26,10 +26,10 @@
     "type.union.GetResponse7": "Type.Union.get.Response.anonymous",
     "type.union.GetResponse8": "Type.Union.get.Response.anonymous",
     "type.union.GetResponse9": "Type.Union.get.Response.anonymous",
+    "type.union.GetResponseProp": "Type.Union.get.Response.prop.anonymous",
     "type.union.GetResponseProp1": "Type.Union.get.Response.prop.anonymous",
     "type.union.GetResponseProp2": "Type.Union.get.Response.prop.anonymous",
     "type.union.GetResponseProp3": "Type.Union.get.Response.prop.anonymous",
-    "type.union.GetResponseProp4": "Type.Union.get.Response.prop.anonymous",
     "type.union.IntsOnlyClient": "Type.Union.IntsOnly",
     "type.union.IntsOnlyClient.get": "Type.Union.IntsOnly.get",
     "type.union.IntsOnlyClient.getWithResponse": "Type.Union.IntsOnly.get",
